Remove natural key from model and add new fixture dump command
Database dumps should happen with `--natural-foreign` to be able to
refer to automatically generated models such as `auth.Permission` or
`contenttypes.contenttype`.

If our own models include natural keys, django will try to refer to them
by that as well. We currently don't have any option to load the model
from the natural key however and reloading the fixture will fail because
of that (it actually will complain that the keys are not integers).

This is why I removed the natural keys from our models. For more
information about natural keys in case you want to add them again see
https://docs.djangoproject.com/en/1.10/topics/serialization/