Make Question and GivenAnswer serializable again
First off, the model was wrong. There was a AnswerSlider class that
inherited from Answer (which contains just the name of an answer to a
Question) which had a rating field that ranged from 1 to 10 implying
that it would have a connection to a user who answered the question. The
right parent for that class would have been GivenAnswers and so it was
moved to GivenSliderAnswer.

Additionally @m.chlepakov insisted that we supply the ranges for the
SliderQuestions from the database instead of specifying that all our
sliders range from one to ten which is why the SliderQuestion submodel
was added which does exactly that.

Unfortunately, the django_rest_framework provides no way to dynamically
serialize submodels. I had to hack together my own solution in
app/utils.py. It works reasonably well but offers no validation. Fields
are just added to the result as python sees them.

In the future, this should be extended to track the submodel fields as
real rest_framework.serializers.Field objects like the other model
fields