Don't use nondeterministic functions in app.models
In this case it is `os.path.join("product_type", "thumbnail")` which
returns a different result (which results in a different migration)
depending on whether `python manage.py makemigrations` is run on Linux
or on Windows.

It is also redundant to use os.path.join here since the field value is
passed to the save() method of Djangos Storage provider which expects a
path with unix style path separators (as implicitly described
[here](https://docs.djangoproject.com/en/1.10/ref/models/fields/#django.db.models.FileField.upload_to)
("upload_to may also be a callable, such as a function. This will be
called to obtain the upload path, including the filename. This callable
must accept two arguments and return a Unix-style path (with forward
slashes) to be passed along to the storage system.")