Turn Suggestions into a ListAPIView
This means that we will have access to the ListModelMixin from drf that
provides a list method that returns a response containing a paginated
view of the queryset (pagination does not work as of this commit because
the paginator is not set. This will be implemented some time during the
next few commits).

This means that we don't have to implement get ourselves (the default
implementation just calls list)