Make Suggestions.post() work correctly
Currently it calls get_paginated_response which isn't really a method
that you are supposed to be calling if you don't know exactly what you
are doing. The current code calls it with only two of three required
parameters.

A better option is to use the list() method provided by the
ListModelMixin of ListAPIView which proivdes a paginated view of the
queryset