Implement a static version of our Swagger
It's a shame that coreapi (the schema format that Django Rest Framework
uses) can not describe our API properly. The Changes I made to the
generated schema come entirely from information that could have been
extracted from the SuggestionsInputSerializer and I would have loved to
submit a patch to DRF if it would have been possible to implement it
without a design decision on the maintainers part.