Move SuggestionsInput deserialization to get_queryset
The reason for this is that django cannot store objects in the session
(it can save them but restoring them returns a list instead of an
object).

After this commit, post() will just store the input dict in the session
and the validation will happen when processing the request by evaluating
the queryset.

Additionally the validation logic is changed to raise APIExceptions
instead of returning error responses through a check.