Implement cached decorator and use it throughout matching
I also removed the explicit expiration time from the cache calls. The
rationale for it is, that if we have only one default cache timeout, we
might as well set it globally in advisor/settings.py. The default
timeout of 300 seconds is also suitable IMO but I'll be glad to change
it if we feel a particular need to do it.