Fix Scenario.save() rating fixup
values_list('category') returns a list of tuples like [
(1,), (4,), ...]. Taking the first one of this list would result in a
tuple that only contains the id of the first rated category. This in
turn would mean that the code would create a new rating for every
category except the first one that the scenario is rated in which is
obviously wrong.

The fix uses values_list correctly (by getting a flat list with
flat=True) and gets the list of unrated scenarios with a django filter
query instead of a manual set comprehension.