Remove SetUp method from SchemaTest
The generated schema is nice but I don't think it is going to be used
much outside of the current test case. The other tests that are going to
be added to the class are going to use the swagger schema which is going
to be added in a future commit.