Make swagger defaults more explicit
The required properties product_preference and renovation_preference of the
matching input did not have default values because the swagger ui autogenerated
them from the available range (the first one is an enum -> first element, the
second one is a boolean -> true).

The unit test that I am currently writing however requires that the defaults are
explicit which is why I am fixing this inconsitency now.