Test swagger default inputs by validating them ...
against the schema that is given in swagger. You might think that this
is redundant if the swagger schema itself is valid but the specification
actually places no restrictions on default values and you could put the
string "horseshit" in every single default value and no swagger
validator would care.