Promote SchemaTest to management command
It can now be run with `python manage.py testapi`. The reason for this
change is that `python manage.py test` tests run against a separate test
database which would not correctly emulate what the frontend sees when
they send requests to our production server.

I hadd to add `testserver` to the `ALLOWED_HOSTS` because the django
test api client sets that host header by default. It's not a big deal
security wise as django only uses this to prevent csrf attacks that are
caused by url construction and `testserver` isn't a routable dns name.