Split Suggestions and ProductList serializers
Previously, the ProductList was requiring a few parameters that don't
make any sense there (scenario_preference, product_type_filter and
subcategory_filter, all of which only affect the scenarios that may be
suggested and have no effect on the implementation of the shopping
bakset). This was because it used the same serializer as the Suggestions
endpoint.

This commit splits the serializers to match the different requirements
of both endpoints better. (Almost all, more on that later) common fields
have been refactored to `serializers.MatchingSerializerBase`.
Unfortunately, both endpoints have different requirements of the
`shopping_basket` field (the product list requires it to be non empty)
which leads to some code duplication.

For positive news: impementing the shopping basket does not touch any of
the suggestions-phase-only fields of the user preferences which means
that this change does not affect the matching *at all*. Yay for weak
typing!

This is going to need some serious refactoring and possibly some
rebasing later to deal with the code duplication. I'm commiting it now
because the number of changes is getting kinda large and I want to
deliver a mock of the product selection endpointy by thursday evening.

Another thing that has to come later are checks and comments to the
things in the logic module to ensure and make it clear that
`scenario_preference`, `product_type_filter` and `subcategory_filter`
are not touched unless there is a suggested scenario on which they can
be applied.