[Refactoring] Remove code duplication
By creating a MatchingInputBase namedtuple that contains everything that
is not exclusive to either phase of the consultation process.

This includes the shopping basket since only the requirements on the
field but not the actual serialization change. The default is that the
field is optional. The ProductListSerializer actively overrides it to
make it required.