Mock replacement slots.
They take the meta device id of the broker of a random scenario that the
product is used for. Most certainly wrong but also most certainly an
input that is a valid slot. If the frontend tries to replace that slot
they will probably get an error. We don't care about that.