removed originals_from_merged frunction
it is not needed anymore. The only place where it was needed (and
also cased the key error bug) changed in behavior. All scenarios
that are needed for validation are stored inside the deivce mapping
and can simply read from this dict. No need to find the the orignal
meta device anymore. (it was also miss implemented and shouldn't worked
in the frist place, don't know how it worked)

TL;DR
However the bug was cased by acessing a orignal meta device (to find
the scenario of this device) that was merged into an other meta device
and so removed from the device list.
The searching for the original meta device is not needed becuase the merged
one contains all scenarios that its subtituded meta device belonged to