Retain information about meta device -> product assignements
... in compute_matching_product_set.

It works by replacing the way paths from a endpoint implementation to
the master broker are stored internally. Previously a path was just a
set consisting of all involved products. Now it is is a custom object
which stores the previously mentioned set along with the product that
implements the meta endpoint.

Since the paths are stored in a meta endpoint -> path set dictionary,
this will allow us to extract replacement slot information out of this.

I also had to redo the app.utils.dict_cross_product method to lift the
information on which meta device pulled in a certain path out of the
inner loop.

The new dict_cross_product takes a dictionary mapping things to sets of
other things and returns an iterable of dictionaries that map each of
the original keys to one choice out of their choice buckets.

Previously, the information what key the choice originally came from was
thrown away and we just got the union of all the chosen paths.